[FE][Refactor] constant 변수, 타입 분리
- Appconfig 파일 경로 변경
