+++ frontend/src/lib/constants/commonConstants.ts
